+++ src/main/java/de/samply/teiler/response/entity/CreateQueryResponseEntity.java
@@ -1,0 +1,7 @@
+package de.samply.teiler.response.entity;
+
+public record CreateQueryResponseEntity(
+    Long queryId
+) {
+
+}