+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.0.0 - 2022-12-09]
+## [1.0.0 - 2023-01-10]
 ### Added
 - Search Controller
 - Blaze Store Client
@@ -40,10 +40,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Java 19
 - Template in request body
 - Attribute condition fhir path
+- Teiler DB
 - Attribute join fhir path
 - Attribute condition id and value fhir path
 - Excel Format
+- Flyway config
+- Query entity
+- JPA config (Hibernate)
+- Get all queries (paging)
+- Query Execution and Query Execution File
+- Download files (also as zip)
+- Clean Temp and Write Files Jobs
+- GET active, error and archived inquiries
+- Format inquiry timestamp
+- Allow Cors
+- Default value for inquiry label and decription
 
 ### Changed
 - Rename Conversion Template as Converter Template
 - First anonym: 1
+- Cors config